+++ src/main/java/ru/made/ex/one/TradeFactory.java
@@ -1,7 +1,7 @@
 package ru.made.ex.one;
 
 class TradeFactory {
-    static Trade getTrade(String tradeType, float price) {
+    static Trade getTrade(String tradeType, double price) {
         return TradeType.valueOf(tradeType).getInstance(price);
     }
 }
